refactor(session): extract expression evaluation into src/session/inspection/
`evaluateExpression`, the redaction name rule it depends on, and the
`EvaluateResult` shape move to `ExpressionEvaluator`. `EvaluateResult` is
re-exported from `session-manager-operations.ts` so its two existing import
paths keep working.
